src/vscode_marketplace/requirements.txt
django-cors-headers
djangorestframework
markdown
django-filter 